Optionally force compilation with MacFUSE.
Use known MacFUSE compiler and linker options instead of pkg-config when
building with NOPKGCONFIG. An official build should compile with MacFUSE
to make sure the application will work regardless of whether an end-user's
system is using MacFUSE or its successor OSXFUSE (with MacFUSE
compatibility).
Not using pkg-config also has the advantage of not having to install
pkg-config (through e.g. MacPorts, Fink, or Homebrew) which makes it
easier to do a build on an offline system.