Made replacement Makefiles
The old Makefiles needed to be heavily modified in order to achieve
deterministic builds and cross compilation, among other things. I
thought it would be more productive to start from scratch than to try
to "fix" the old Makefiles.

While this fully builds Linux targets, overall, it's a work in
progress.

Improvements include:
	- Replace recursive make with includes
	- More target templates
	- Dynamic output directories
	- Cross compile preparation

Yet to be re-inplemented:
	- Per-component root Makefile (make -C Platform,
	 for example)
	- Fully compile for targets other than Linux