Example program that works in VMWare
VMWare Workstation Pro 12.5.7 (Windows and Linux)

Jason wrote this while we were last debugging our bootloader in
VMWare. Our bootloader functions in QEMU (with and without KVM
acceleration) but does not function in VMWare (with and without any
of the acceleration options). The theory is that our invocations of
gcc are producing assembly that is invalid in real mode, and VMWare is
enforcing the rules while QEMU is a little more flexible about it.

This test program is simply to act as *something* we can build that
functions in VMWare. It can be injected into the build process with:

make clean; cat main-loop.s > main.s; make main.vmdk