Eliminate extra ld invocation
For organizational purposes, I used bootcode.ld to link the generic
payload of the boot sector (after organizing contents), and used
mbr.ld and fat32.ld to plant the pre-organized boot code and
necessary jump instructions into the sector accordingly. This made
more sense when I was outputting in binary format, so bootcode.o's
symbols could be investigated with objdump where boot.com cannot. Now
that boot.o is being created, and offers the same benefit, the
intermediary ld is just unnecessary and appears to be one of two
issues with building with Cygwin's gccs.

The other issue with Cygwin is that their (mingw's?) ld is not
configured to support any kind of object file; it can only target PE
executables.