Add single-section linker example
I've looked a little further into the PE issue described in
19ca3d32e11cc09048186df70777b43c787824d0.

Some things were incorrect; specifically, that formats other than PE
have a have any sort of direct correlation between VMA and file
offset. In fact, it appears that the issue I encountered is generic
to all the executable formats: they expect that an OS will relocate
the sections.

It appears that the only reason the test succeeds with ELF, A.OUT,
and COFF, is because ld will consecutively output consecutive
sections. This appears to be the expectation of image formats used by
a multiboot kernel (at least for early code) [1]. If they are not
consecutive by VMA, they may not be consecutive in the file. The
difference with PE is that all sections are put into a separate (4K)
page, even if their VMA/RVA says that they should be consecutive in
memory. As a sidenote: I was unable to make PE sections output
consecutively even when 4K page alignment would encourage it (a .text
section padded to exactly 1 page size was output with an empty page
between .text and .data).

Even though this works well with ELF, A.OUT, and COFF, it should be a
general rule that sections being consecutive in the image file is not
guaranteed. However, addresses relative to a single section do seem
to be guaranteed by all object formats.

In fact, the NetBSD kernel offers an interesting example of how
things are done by real kernels. The comment for
multiboot_pre_reloc() in src/sys/arch/i386/i386/multiboot.c of NetBSD
7.1.1 says:
WARNING: Because the kernel has not yet relocated itself to KERNBASE,
special care has to be taken when accessing memory because absolute
addresses (referring to kernel symbols) do not work.  So:

    1) Avoid jumps to absolute addresses (such as gotos and switches).
    2) To access global variables use their physical address, which
       can be obtained using the RELOC macro.

In NetBSD, relocations are done according to parsing the ELF of their
kernel. While this is not the same issue as section VMA differences
not matching file offset differences, it shows that their practice is
to regard the early boot code as layout correction before starting
normal execution.

Considering this, I think multi-section code is "unsupportable"
without similar functions to parse the image format.

This also makes the usage of _edata and _end interesting. They imply
that the loaded code is flat binary (for which consecutive/singular
sections effectively make). When using image parsing techniques, the
whole image should be loaded, and these fields should be set to 0.

For our purposes, the alternative to parsing our image and correcing
layout is to rely on the addressing guarantees of single-section
linking, or formats that can output sections consecutively. This
still offers advantages over real flat binaries such as symbol tables
and other debugging features.

[1]
"load_end_addr - Contains the physical address of the end of the data
segment. (load_end_addr - load_addr) specifies how much data to load.
This implies that the text and data segments must be consecutive in
the OS image; this is true for existing a.out executable formats."
See: https://www.gnu.org/software/grub/manual/multiboot/multiboot.html#Header-address-fields
See: https://web.archive.org/web/20180210080447/https://www.gnu.org/software/grub/manual/multiboot/multiboot.html#Header-address-fields