Workaround ld+objcopy issue regarding section flags
objcopy is refusing to copy the .eos section because it doesn't have
a flag that marks this as appropriate (see objdump -h output on
boot.o without this patch applied). ld sets these flags based on the
input sections, and doesn't appear to provide a *generic* way to set
these flags to the appropriate target-specific values.

So... working around by providing an empty section to copy flags from
(in the form of a compiled empty assembly file).