Basics for unit tests
The basic idea is to put unit tests in a kernel and running them in
QEMU.

This was appealing because QEMU can load Multiboot kernels directly,
and the kernels can be of arbitrary size (until 16-bit segment size
becomes an issue). QEMU also supports an ISA debug device that, when
written to, will immediately exit QEMU with the specified return code.

With QEMU doing all of the loading work, our kernel's implementation
can be limited to simply doing the tests and can easily express
success to the host.

Of the two linker scripts, I am using the unisectional one by default
to work around an issue with consecutive sections not guaranteed to
be consecutive in PE images.
See commit message of: 2a19f6d17e42323ddde1c595ebcb5e5d11d770c6