Fix ELF entry and remove leading jmp
GNU's Multiboot example starts with a jmp to the entry point just
before the Multiboot header. In inspecting this, I found that this
appeared to be necessary for ELF, because the other formats got their
entry point from the Multiboot header.

Well, I just realized that Multiboot loaders are parsing the ELF
header for entry point, so the jmp should be unnecessary in both
cases. I think GNU includes the jmp because they never use ENTRY() in
a linker script (so ELF's entry defaulted to first instruction).