Create lib module for common code
The naming may not reflect this perfectly, but src/lib is where the bulk
of our functionality will exist.

The intention is for the one copy of the code to be compiled from
src/lib, and is linked to src/main and src/test, while src/main and
src/test use that same code under their different contexts.

At the same time, this is also going to be an actual library of code
that is common between the modules. For example, the plan right now is
to make an MBR that loads the bulk of the bootloader code from a PE
section, and I expect I'll use the same functions between these two
artifacts.