Add drives to unit test
In preparation of I/O development.

Note on the "test_run: bindir:=${bindir}" line: While Make immediately
expands variables for targets and dependencies, it does not expand
variables in recipes until the recipes are called. Because of this,
"${bindir}" in the kernel recipe was expanding to a path derived from
one of the "${MDEPS}" Makefiles, rather than the current Makefile.

By using ":=" to re-set the the variable for that target, I'm
effectively pre-expanding it at target declaration, so its value at
run-time of the recipe matches the value used to define the target in
the first place.